refactor(polynomials): share rational-function arithmetic bounds
